requirements.txt
openai==1.12.0
tweepy==4.10.1
schedule==1.2.0
python-dotenv==1.0.0